# GS-hw: Optimize blending equations based on alpha value.