@lightningterror 2022-02-22
Sly 2 pr 128 fps , master 115 fps - OpenGL
Vulkan and dx remain around the same. I assume for amd it will help on Vulkan too.