@tellowkrinkle 2022-02-22
Instead of flagging specific entries with another flag, apply this rule: "If C is 128 && D == B then swap D and A".  This will rely on our existing handling of the A == B case to also handle D == B cases when B is known to be multiplied by 1.

Also for future changes, C being 128 removes the chance of HW blend rounding issues, so you can use HW blend in Ultra

We should also work on normalization to reduce the number of different pipelines
If A == B then set A, B, and C all to zero