@lightningterror 2022-03-04
> We should also work on normalization to reduce the number of different pipelines
> If A == B then set A, B, and C all to zero
Yeah, it was on my mind already.

> Instead of flagging specific entries with another flag, apply this rule: "If C is 128 && D == B then swap D and A". This will rely on our existing handling of the A == B case to also handle D == B cases when B is known to be multiplied by 1.
Yes but I have some other ideas so I might still go with a flag, well one flag instead of 2 like in this pr but changing blend mix 1 to 4 for it.
https://github.com/PCSX2/pcsx2/compare/blend-work12