@lightningterror 2022-03-16
Benchmark comparison eyecandy:

```
Getaway/Getaway Black Monday
opengl, ultra blending:
pr: 26 fps, 5720 draw calls
master: 13 fps, 13507 draw calls

vulkan, ultra blending:
pr: 49 fps, 5720 draw calls
master: 41 fps, 13507 draw calls
```

SOTC
```
opengl, ultra blending:
pr: 41 fps, 7143 draw calls
master: 20 fps, 17693 draw calls

vulkan, ultra blending:
pr: 91 fps, 7143 draw calls
master: 63 fps, 17693 draw calls
```

Other games are also affected just to a lesser extent but still show an improvement, some big, some small, some no change.